Bug fix for the iterative solver of stomata conductance
The original solver occasionaly leads to negative ci. This revision fixes
this bug. Also, it was found that the old gs-An coupling
did not guarantee a solution for all situations. For cases without solution, the revision
forced the solution to converge with gs at this minimum value. A comparision between
orginal and the fixed versions for 1850 SP simulations suggest that the difference is
small, with a slight cooling in the tropics.