Subcycle hydro dt under master model dt
This expands the adaptive time step capability of the hydro model to
subcycle the hydro model until the master model dt is reached.  This
allows for the vastly smaller time steps required by the hydro model.
It also solves two issue related to the old implementation:

1. time steps of less than a second (which commonly occur in the hydro
model) were not being proprely handled by the timekeeping, leading to
very large errors in elapsed time.  (E.g., the hydro model would set the
dt to 1e-2 seconds, yet the clock would advance 7 seconds.)  With this
change, the timekeeping only sees the 'reasonable' time steps of the
master model, avoiding this issue.

2. The requirement of millions or billions of hydro time steps led to
enormous log files (many GB) as messages related to the hydro model and
messages from the rest of the LI model were being written over and over.
The new implementation only writes messages at the frequency of the
master model time step, vastly reducing the size of the log files.
Presumably performance is also improved, but I have not checked this.

Note that I have not checked the model bevaior when not using the
adaptive timestepper in the hydro model, but I think it should still
work.  However, I do not see any practical utility to using a
non-adpative time stepper for the hydro model besides some basic
testing, so it may make sense to make the adaptive timestepper the only
timestepping option for the hydro model.