Avoid setting invalid pointers to scalar constituents in mpas_atmphys_interface module
Several scalar constituents, e.g., 'ni' and 'nr', are only allocated by specific
microphysics schemes through the use of packages. In the routines microphysics_to_MPAS
and microphysics_from_MPAS, we should only set pointers to sub-arrays of the scalars
var_array for constituents that are guaranteed to be allocated based on the choice
of microphysics scheme.

Thanks to Wei Huang for finding this problem in MPAS v5.0.