Changes to support MISMIP+ experiments with a GLP
I previously ran the MISMIP+ experiments on a branch created from landice/develop.
I am now repeating them on a branch based on landice/friction_dissipation_and heating.
This branch has interfaces linking to a newer version of Albany containing the GLP.
I rebased the branch onto the latest landice/develop, and with this commit I am adding
the following changes:

mpas_li_thermal.F: I changed some logic so that the model does not immediately return
when config_thermal_solver = 'none'. The reason for this is that MISMIP+ does not
need a thermal solve, but it needs basal melt rates that are currently computed in the
thermal module.  A better long-term solution (to be implemented later) is to move the
basal melt calculation to another module.

mpas_li_advection.F: I added a call to li_calculate_mask before the surface and basal
mass balances are applied. This is necessary to ensure that basal melt rates are applied
to cells that were previously ice-free, but now contain a small amount of floating ice
due to advection.

mpas_li_diagnostic_solve.F: Per Matt Hoffman's instructions for the new GLP, I changed
the computation of betaSolve.

albany_input.xml: I changed an Albany setting to activate the GLP. This will likely be
the default setting for MISMIP+.

I do not expect any changes for standard tests, but the new call in the advection module
means that this commit is not BFB when basal melt rates are applied.