Add ESMF_TimeIntervalProdI8 routine to multiply a time interval by an 8-byte int
To complement the existing routine ESMF_TimeIntervalProdI that implements
the '*' operator for multiplying an ESMF_TimeInterval by a 4-byte integer,
this commit adds a new routine, ESMF_TimeIntervalProdI8, that multiplies
an ESMF_TimeInterval by an 8-byte integer. This new routine is accessed
through the '*' operator in the same way as the existing
ESMF_TimeIntervalProdI routine.

This commit mirrors commit 656670e in CIME.