Add mpas_abort Makefile dependencies
When the mpas_abort module was added recently, no explicit dependencies
were added to the framework Makefile.  This caused parallel builds to
fail on some processor counts (in my experience, 3).

This commits adds the dependencies for the mpas_abort module to the
framework Makefile.  This fixes the -j 3 parallel build.