Provide compatibility with ECMWF 100-289 cm soil layer
This commit introduces changes to allow the init_atmosphere core to properly
handle the ECMWF 100-289 cm soil layer (for both soil temperature and soil
moisture).

In older versions of the WPS, the 100-289 cm layer was decoded as spanning
the depths 100-255 cm; however, the actual bottom depth of this layer in
ECMWF is 289 cm, and this was later corrected in the WPS. In order to be
compatible with this new, correct layer, additional block of code are added
in the init_atm_case_gfs( ) routine.

At present, compatibility with the 100-255 cm layer is maintained, and
the 100-289 cm layer is treated in a completely independent way.