Update to bless_test_results
There's no need to abort the namelist bless if there's not a namelist
phase, just do the bless.

Both namelist and baseline blessing should take into account the state
of the corresponding phase. If the corresponding phase is PASS, we can
assume there's no need to bless unless the user passes a special new
-p option.