Bring back a commented-out line in the circle_area CMakeLists.txt
Jim Edwards commented out this line, but I think it should not have
been commented out. This test is currently failing with and without
this line. I'm guessing he commented it out thinking it would make the
test pass, but it doesn't.

Also add a README noting the expected failure