Hook up log module to framework, test core
This commit integrates the new log module into the framework and sets it
up for the test core.  This is done by:

* Adding an instance of mpas_log_type to the domain type.  Thus every
domain instance will now have a log manager instance, whether or not it
is used.
* Adding a setup_log function pointer to the core type.  This means that
each core will need define a setup_log function at which to point.
* Having the subdriver call the core's setup_log routine at the earliest
possible opportunity (so that subsequent init steps can make use of it
for writing log messages!).
* This commit adds {CORE}_setup_log routines to all cores.
This function calls the log module's init routine to set up the log manager
for this core and store a pointer to it in the log module.
For now I've added a line that disables the log file from being created.
This can be removed when each core is ready to start using the log
module.

As of this commit, the log module actually creates a log file and writes
a message in the test core!

Issue to be aware of with this commit:

* As of this commit, the new log file is created in every run (if it is
enabled - currently only in test core), currently with the file name
'stick.0000.out'.  ('stick' is an under-developed 'log')