Separate log init and opening of output log file
To make the log manager customizable to both a core and an external
driver (e.g. climate model), a better design is to separate the
initialization of the log manager from the opening of the output log
file.  By making these separate routines, the core and/or driver has an
opportunity to modify the default configuration before the output log
file is actually opened.  This commit reorganizes operations to make
that separation.

It also extends the newly named mpas_log_open routine to work for either
an output log or an error log.  Thus this same routine can be called by
mpas_log_write if an error message is presented and it comes time to
open the error log.  This adds some slightly convoluted logic but allows
code re-use.  Note that there is a distinction that the core/driver can
choose to disable any output log they would like, but there is no option
to disable an error log - it is inherently assumed that we never want to
disable the generation of error log files by any task if that task
presents an error message to mpas_log_write without the masterOnly flag.