Add error message handling to mpas_log_write
This commit adds the logic to handle error and critical error behavior
properly in mpas_log_write.

An error message causes the error log to be opened if it is not already.
Error and critical error messages are written both to the output log and
the error log.  This occurs regardless of debug or optimized mode.
However, there is the option to exlude errors (but not crtiical errors)
from all but the master task.  If a critical error is presented, then the
model also aborts.

This commit also adds a timestamp to the opening and closing of each
file.