Conditionally use undefine in Makefile for backward compatibility
The Makefile undefine directive was included in make 3.82 . Many
machines (e.g. blues, compute001,..) still use make 3.81 . This
fix uses the "undefine" directive only if the variable is already
defined (On machines that use make 3.81 we are careful not to
define the variable, PNETCDF_PATH, when building with mpi-serial).

Fixes #1033