Add namelist option config_extrap_airtemp for extrapolation above/below first-guess levels
This commit adds a new namelist option, config_extrap_airtemp, in the new
namelist record &interpolation_control to specify the method by which air
temperature is extrapolated beyond the first-guess levels.

Note that this commit only defines the namelist option in the Registry.xml file,
but does not actually add code to make use of this new option.