Add routine to expand variable values in a log message string
This routine allows the expansion of variable placeholders indicated
with $ into a formatted version of a variable value passed in through
optional arguments.  The routine is documented with usage details.
The exp formatting is not currently implemented, and the real formatting
may need to change in the future, but the routine functions.
Design is based off of mpas_expand_string.

It has been hooked up to mpas_log_write so all incoming messages get
formatted.