Revise log_expand_string to use fill value, detect precision
This commit revises log_expand_string in two ways:
1. Inserts a fill value of '**' of the number of expansion indicators
is greater than the size of the variable array for that data type OR if
an invalid expansion indicator is presented.
2. Checks the precision of real variable values and adjusts the format
accordingly.  Note I think that this could also be accomplished with an
ifdef, but it seems preferable to avoid preprocesser directives if we
can handle the operation in-code instead.