Assign file units on init, add mpas_log_file_type
This commit fixes a design flaw when using externally supplied file
units.  In the case of the error log, we do no necessarily want a
pre-opened file unit, but instead want to be able to open the error log
later in the simulation if an error occurs.  The design flaw is that if
we do not 'reserve' a file unit from the driver (e.g. climate model),
and instead let mpas choose the file unit for the error log file, we run
the risk of clobbering an existing open file used elsewhere in the
calling code (e.g. climate model).  (This is because mpas_new_unit does
not check if a unit is in use, but instead relies on an internal table
of units that MPAS itself has assigned.)

The solution in this commit is to assign the unit number for both the
output and error log files on init.  For the output file, the unit is
still opened on init.  The the error file, this unit number is just a
'reservation' for later if/when the error file needs to be opened.  By
doing this at init, it allows the calling system (e.g. climate model)
the opportunity to supply this unit (and track in-use units on its own).

In implementing this change, I found it much less confusing to expand
the mpas_log_type derived type by moving the attributes specific to
output and error logs into two instances of a new mpas_log_file_type
derived type.  This makes the code easier to read and allow
generalization of some sections.

These changes required a minor update to the *_setup_log routine in each
core, so those changes are also in this commit.  As an added bonus, with
the new file data structure, I am able to truly disable the log module
in each core other than the test core.  (Before an attempt to write an
error message with mpas_log_write would have still opened stick.*.err.)