Fix bug in ERP namelist comparison.
The following pattern was broken...
% ./create_test -o -g ERP.f45_g37_rx1.A.melvin_gnu -b jimtest
% ./create_test -c -n ERP.f45_g37_rx1.A.melvin_gnu

In the first command, the cmpgen_namelists occurs during the RUN
phase and therefore occurs after the ERP build phase which impacts
the contents of the namelist files. In the second command, the cmpgen_namelists
occurs after the setup phase and therefore will not match the first.

With this change, we have essentially re-introduced all the namelist
creations that I tried to remove a few months ago in order to
improve performance. The SETUP phase is back up to 15 seconds on melvin.