Refactor env_batch
In places all over the code, env_batch was treated like a standard
env*.xml file, but it's format was not compatible due to the way
jobs were being handled. This required special code all over the place
to handle and it also rendered xmlquery useless when querying for
stuff in env_batch.

This branch fixes that by re-doing the way jobs are expressed in env_batch.xml
by treating them as groups instead of sub elements of the useless job_submission
group.

This also fixes a bug entry_id.get_value where it was throwing an error
complaining of too many matches even though the user provided a subgroup
that would have limited it to a single match.