Introduced scaling of the gravity-wave absorbing layer coefficient
This commit modifies the formulation of the upper gravity-wave absorbing layer
to be scale-aware by multiplying the damping coefficient (config_xnutr) by
dx/dx_fine (dx_fine is the mesh spacing within the highest-resolution region of
the variable resolution mesh) when setting the damping layer configuration at
the beginning of an MPAS integration.  The mesh density function used to
generate the mesh is also used to determine this ratio.  Analysis of
the gravity wave absorbing layer characteristics and test results indicate
that this scaling is needed for the absorbing layer to effectively damp
vertically propagating waves in the coarse-resolution region of
variable-resolution meshes.