Update runoff_to_ocean build process
The configure script has been updated so the directions in INSTALL were a
little out of date. (This will need to be updated again after issue #1265 is
closed.) Also updated the clean target to also remove executable (removing need
for realclean target) and the distclean target to clean up after configure.