Add missing uses of 'mpas_io_units' in atmosphere core to enable compilation
Several modules in the atmosphere core did not use the 'mpas_io_units' module
to obtain stderrUnit, presumably because this variable was inherited through
some other framework modules that were used. Now that the framework has been
converted to the new mpas_log module, we need to temporarily use mpas_io_units
to allow the atmosphere core to compile.