Allow for logging of strings with literal '$' characters
The mpas_log_write( ) routine expands variables, designated as $r, $i, and $l,
using the '$' character as an indicator of a variable to be expanded.
This commit enables a literal '$' character to be printed by adding '$$' in
the string, e.g.,

 call mpas_log_write('Use $$r to insert a real value')

now prints to the log file the following:

 Use $r to insert a real value

This commit also includes a public helper routine named mpas_log_escape_dollars
in the mpas_log module. The mpas_log_escape_dollars function takes as its only
argument a string and returns a string constructed by replacing all '$'
characters in the input with '$$'. The returned string is of the correct length,
i.e., there is no need to wrap the return value with trim( ).