Don't define PNETCDF_PATH for mpi-serial+anvil
Making sure that we don't define PNETCDF_PATH when using mpi-serial
on Anvil.

Without the fix if PNETCDF_PATH is defined for mpi-serial builds
it gets "undefine"d in the Makefile. However older versions of
the make utility do not recognize the "undefine" directive. This
causes mpi-serial builds to fail.

Fixes #1340