Pass pressure-corrected temperature to Albany
While verifying MPAS LI thermal calculations, I found that the heat dissipation computed
by Albany was too small.  There were two reasons for this:
(1) A factor of 2 was missing in the heat dissipation calculation in Albany.
(2) Albany was computing the flow parameter A using the Paterson-Budd relationship
with the true temperature instead of the pressure-corrected temperature (i.e., the
temperature relative to the pressure melting point, plus 273.15). This made the
flow factor too small (by a factor of ~1.5) and the ice too stiff.

For debugging, I inserted the missing factor of 2 by hand in mpas_li_velocity_external.F.
This correction is now commented out, anticipating that the factor of 2 will be present
in future versions of Albany.
The commented-out line can be deleted once we verify it is not needed.

Also in mpas_li_velocity_external.F, I computed the corrected temperature
as a function of the difference between the true temperature and the pmp temperature.
The corrected temperature is now passed to Albany instead of the true temperature.
Mauro confirmed that the temperature (corrected or not) does not appear elsewhere
in Albany.

With these changes, the interior heat dissipation and basal frictional heating
computed by MPAS LI agree (to within a few percent) of the values computed by CISM
with a similar FO solver and with similar values of ice thickness and temperature.
I verified this by running EISMINT-2 experiments A and G with CISM, then interpolating
the spun-up thickness and temperature to the MPAS grid and running the same experiments
in MPAS LI for a few timesteps.

Answers will change for any problem using the Albany FO dycore with a prognostic
temperature.