Initialize horizontal fluxes for ghost-cell faces in monotonic transport
In the atm_advance_scalars_mono_work( ) routine, the flux_arr field holds
the horizontal fluxes through cell faces. This array was previously only
set for edges that bordered owned cells, leaving the fluxes between ghost
cells uninitialized. Later in the routine, however, values from the flux_arr
field are used for all edges, and this can lead to floating-point exceptions.

The fix in this commit simply sets the flux_arr field to zero for edges that
separate two ghost cells in the loop over edges.