For PIO2, set 'pio_mode' to zero when creating netCDF-4 files
Internally, when creating a netCDF-4 file, PIO will perform a logical OR
of the file mode with NC_MPIIO and NC_NETCDF4. Previously in MPAS, we
set the file mode argument to PIO_64BIT_OFFSET (which maps to NC_64BIT_OFFSET),
and apparently the logical conjunction of all three of these flags
(NC_MPIIO, NC_NETCDF4, and NC_64BIT_OFFSET) is not valid when creating
a parallel netCDF-4 file. The proper approach with PIO2 is to simply pass
a null (zero) file mode when we are creating a netCDF-4 file.