add varibles to send burnt fraction from alm to betr
Because when betr bgc is active, a different som pool
structure will be used, such that the existing fire area
calculation is no longer legitimate. This is fixed by
adding new cwd variables and fire fluxes is done in betr.
Also added a few other modifications to ensure consistency
in mass budget.