Zero the x2g fields if .not. glcrun_avg_alarm
In principle we shouldn't need to zero the fields at these times -
instead, glc should just ignore the fields at these times. However, some
tests (like an ERS or ERI test that stops the final run segment
mid-year) can fail if we don't explicitly zero the fields, because the
x2g fields can then differ upon restart.

The refactored logic in cesm_comp_mod also results in these changes:

(1) This code:

         !----------------------------------------------------
         !| cpl -> glc
         !----------------------------------------------------

         if (iamin_CPLALLGLCID .and. glc_prognostic) then
            call component_exch(glc, flow='x2c', &
                 infodata=infodata, infodata_string='cpl2glc_run', &
                 mpicom_barrier=mpicom_CPLALLGLCID, run_barriers=run_barriers, &
                 timer_barrier='CPL:C2G_BARRIER', timer_comp_exch='CPL:C2G', &
                 timer_map_exch='CPL:c2g_glcx2glcg', timer_infodata_exch='CPL:c2g_infoexch')
         endif

will now only be executed if glcrun_alarm is true. That's what happened
on master, but the previous version of the branch executed it even if
glcrun_alarm was false.

(2) This code:

                  call cesm_comp_barriers(mpicom=mpicom_CPLID, timer='CPL:GLCPREP_BARRIER')
                  call t_drvstartf ('CPL:GLCPREP',cplrun=.true.,barrier=mpicom_CPLID)
                  if (drv_threading) call seq_comm_setnthreads(nthreads_CPLID)

and then:

                  if (drv_threading) call seq_comm_setnthreads(nthreads_GLOID)
                  call t_drvstopf  ('CPL:GLCPREP',cplrun=.true.)

will happen even if .not. glcrun_avg_alarm