Default for tag_name is generalized (fixed typos too)
The default for tag_name only worked for code if the code is in a
directory called ACME.  The default will now work for whatever
directory the code is in (assuming run_acme is run from within the
code directory).

Some typos have also been fixed, including some that stopped the script
from completing.