Refactor BUILD_THREADED to match documentation.
Documentation (and CIME2 behavior) states that BUILD_THREADED is
supposed to represent an override to force the model to be built
with threads even if nthrds is one for all components.

Until this PR, we had been using BUILD_THREADED as a marker to
indicate that we found a component with nthrds > 1. This is not
useful because we already had case.thread_count for that.

This PR restores BUILD_THREADED to its original behavior and adds
some tests to enforce.