Enable physics load balancing for CAM-SE by default
Recent ACME experiments across all target platforms, mesh sizes,
and process and thread counts indicate that performance
is improved (or, at least, not degraded) by using physics load
balancing option 2 with the point-to-point implementation of
the load balancing communication operator (phys_alltoall == 1).
This changes the defaults to these option settings. Other options,
including the original defaults, can be set in user_nl_cam.

This is bit-for-bit compared to not using physics load balancing
if the BFBFLAG in env_run.xml is TRUE, which is the current ACME
default.

[BFB]