Change G0 format to Gw
Some modern compilers still do not support the G0 format specification
used in the log module.  See list here:
http://fortranwiki.org/fortran/show/Fortran+2008+status: (last updated
in April 2017)

Therefore, this commit replaces G0 with a specified width using Gw
format.  The width is chosen to be 7 characters larger than the
precision.
E.g.: -1234567.89012345   G13.6   ->   -0.123457E+07
For the cases where not all those characters are used, having too many
characters is ok, because we use trim on the resulting string.