Add new variable 'initial_time' to model initial conditions and output files
In order to help post-processing software determine the forecast lead time
represented by the data in output files, we now include a new variable,
'initial_time', in the model initial conditions, which is carried through to
restart and other model output files. This field is defined in the Registry.xml
file for both the init_atmosphere and atmosphere cores as

<var name="initial_time" type="text" dimensions="" units="YYYY-MM-DD_hh:mm:ss"
                         description="Model initialization time"/>

and appears in files as

char initial_time(StrLen) ;
	xtime:units = "YYYY-MM-DD_hh:mm:ss" ;
	xtime:long_name = "Model initialization time" ;

In future, this variable may also be useful to allow bit-identical restarts
at any time, rather than at special (e.g., radiation timestep) times.