Add path to pio2 cmake modules for HOMME tests
The latest version of PIO1 requires access to both PIO2 cmake
modules (to get the latest version of cmake modules) and the older
version of cmake modules (for some tests that are no longer
available in the latest version of cmake modules).

So adding paths to the latest cmake modules that are available
with pio2.

Fixes #1411

[BFB]