Update MPAS-O submodule to fix block decomposition error
This in the fix to issue #1467, where different block partitions of
the ocean did not match bit-for-bit between runs.  This PR adds a halo
update within the barotropic subcycling of the ocean split explicit
timestep.  The actual cause of the problem is most likely a few
locations with a two-wide halo, where boundary noise is getting to the
domain. There may be a more elegant fix to halo creation later on, but
this PR provides a fix with the correct result.  The change to the
elegant halo fix would be bit-for-bit with this PR.

This closes #1467