Initial work to dump albany mesh to ascii file
This simply sets up a new namelist option called:
config_write_albany_ascii_mesh
If this is true, on init a new routine is calleed that calls a new C++
function in the interface to write the Albany mesh in ASCII format.
Then the model aborts (does not actually run a simulation).

This commit adds the new options and routines.  The C++ function has an
API, but the needed contents are for now commented as a first draft, and
the ASCII write statements have not yet been written.

This commit also adds new fields to Registry for the needed
observational fields to be used in Albany optimization.  They are
package variables that are only enabled (allocated) if the above config
is true, because they are otherwise unneeded by the model.