change default remap to UKMO for dcmip test 4
For the other dcmip2012 tests, we were already using the UKMO remap
algorithm instead of PPM due to strange behavior of PPM near the
boundaries.

For test4, the remap algorithm is a net energy INCREASE, so there
is probably something wrong with how it treats the boundaries in
the NH case.