Send Albany output to log.albany.nnnn.out file
After the new log system was added, stdout and stderr from the Albany
external velocity solver is dumped to the terminal.

This commit redirects it to a file called log.albany.0000.out file.  As
implemented now, only the master task ever writes this file, and the
other tasks are redirected to /dev/null instead.

Note: it did not appear possible to redirect Albany output to the MPAS
log file without first closing the MPAS log file before Albany took over
and then reopening it afterwards.  This seemed cumbersome and
potentially problematic, and, given the verbosity of Albany output,
putting it a separate file had advantages anyway.

The redirection to this new log file is applied to both stdout and stderr.
It is activated before each call to the Albany routines.  stdout and
sterr are then reset back to their original locations between Albany
solves so that if the code is running in ACME, stdout and stderr are not
hijacked during the rest of the model when other components or the ACME
driver might be expecting them to point elsewhere.

In order to make the new albany log file easier to follow, I insert the
timestep number in the file before each solve.  (I could not write the
model time because that is not known yet during the velocity solve when
using the adaptive timestepper.)  To allow writing the timestep number,
I added a new MPAS model variable to the mesh pool called
'timestepNumber' that is updated on each time step.