Remove comment asking if we want norm=.true.
This relates to the mapping from glc to lnd of Sg_icemask. I'm actually
starting to think that we want norm=.false. here: see #1516. However,
for now I'm keeping it as is because (a) I'd like to give this a bit
more thought and/or talk to Bill Lipscomb before changing it, and (b) it
will take some thought & work to make the necessary changes to
prep_lnd_mod to use norm=.false. But we should revisit this.