Initialize garbage elements for ke_edge and ke_vertex in atm_mpas_init_block
The arrays ke_edge and ke_vertex must be allocated outside the threaded calls
to the routine atm_compute_solve_diagnostics. The garbage elements of these
arrays are initialized to zero when the arrays are allocated in atm_srk3, but
the atm_compute_solve_diagnostics routine is also called from
atm_mpas_init_block, where these arrays are also allocated and the garbage
elements are not initialized; this can cause an otherwise harmless FPE.

This commit simply adds initialization of the garbage element for ke_edge and
ke_vertex after their allocation in atm_mpas_init_block.