Make flushNow by default true in log module
Currently, messages from the log module are buffered by default and a
flush is only forced if the flushNow optional argument is true.  In
practice, it has been found that the buffered output leads to confusion
during debugging, because the log contents cannot be relied on.

This commit makes the flushNow argument true by default (which
reproduces the flush behavior of the old log.*.err files).  This
behavior can still be overridden to suppress flushing by setting the
optional argument to false.  (For example if a large set of messages are
to be written at once.)