Move layerThicknessEdge calculation to advection
This commit moves the calculation of layerThicknessEdge from
diagnostic_solve_after_velocity to a new routine called
calculate_layerThicknessEdge in module li_time_integration_fe.
The new routine is called from within routine advection_solver.
This move is to clean up the location of random bits of code.
With this change, diagnostic_solve_after_velocity has been deleted,
because all it was doing was calculating layerThicknessEdge.
I also moved the convenience routine li_calculate_layerThickness to the
li_setup module so it is generally accessible.