Extract subroutine to do the smb renormalization
This should be bit-for-bit EXCEPT that, for smb renormalization false,
we now push the modified qice_g back into the attribute vector. This
means that we maintain the "preemptive adjustment to qice_g to account
for area differences between CISM and the coupler" even with
smb_renormalize = .false. Bill Lipscomb confirmed that this is the right
thing to do.