Eliminate diagnostic_solve_before_velocity()
This moves the various parts of that routine to different places.
Preparation for SIA velocity solve is moved to the SIA module.
Preparation for external dycore solves is moved to the external velocity
module.
A new routine is created called li_update_geometry and put in
li_advection.  It solves upper and lower thickness and can be called
from anywhere as needed.
li_calculate_flowParamA is left in mpas_li_diagnostic_vars.F
li_calculate_flowParamA is called by the hydro solver in case it has not
been calculated yet.
The routine li_calculate_diagnostic_vars has been left as a shell of its
former self.
Right now it only calculates the PMP temperature as a diagnostic.
The routine will now just be used to calculate truly diagnostic
variables (things that are not
needed by the model and only included for output convenience for the
user).