Move subglacial_hydro module to mode_forward
When the hydro project started, it was put in its own directory because
it wasn't clear how big it would end up being.  However, in this
reorganization, I encountered a circular dependency issue with keeping
it where it was.  So I have moved it to mode_forward.  Since it is only
one file, it makes more sense to put it in there.

I also did some clean up to the mode_forward Makefile while I was in
there.