Add jobmax and jobmin to the Titan's batch queue
Add the jobmin=0 and jobmax=299008 to the Titan's default batch queue in
config_batch.xml. The number of 299008 is total physical cores in Titan.

This addition will let all jobs run in the default batch queue and avoid
the failure caused by putting more than one job into the debug queue
simultaneously as only 1 debug job is allowed in Titan.

This fix is a better solution than that in the reverted PR #1534

[BFB]