Fixup: add routine to calculate beta
This routine is a place for beta to be calculated and saved as
betaSolve.  Right now it just sets betaSolve to beta, but in the future
it can include other methods.

This is a fixup, because without this the HO solver does not work
correctly (betaSolve is 0 everywhere without this).