Convert user_mods_dir to abspath earlier
This is needed for relative paths to be handled correctly. The problem
with the earlier placement was that the working directory had been
changed to be inside the new case directory at that point, which made
os.path.isdir(user_mods_dir) False because it was interpreted as being
relative to the wrong starting point.