Fixed problem in test for CAM_CONFIG_OPTS. Added usr_nl_clm template.
The test for CAM_CONFIG_OPTS stopped run_acme execution because
run_acme uses the 'csh -e' option, and the test returned a non-zero
status when it was missing.  A similiar test using xmlquery had the
same problem.  Hence, the test has been replaced with an xmlquery of
the value of COMP_ATM; if it is 'datm' then the data atmosphere is
being used.  This new test is clearer, and resolves the problem.

This commit also includes a template for modifying usr_nl_clm, for the
convenience of users.