Add option to include effective pressure in beta
This allows a Weertman-type basal friction law that includes effective
pressure, like Tau_b = beta * N * u^(1/3) or similar.
Note that the appropriate basal friction law still needs to be chosen in
Albany.  This options simply multiplies the beta field that was input by
the effective pressure in the model.

This commit also adds a pointer to LI packages in the li_setup module,
similar to liConfigs.  This facilitates a check to be sure the SGH
variables are allocated before attempting to get effectivePressure.