Add fix for "grounded" boundary nodes below sea level
When the FEM mesh is created, the boundaries are formed by a buffer of
one ice-free node on all sides.  This is so that the last MPAS cells
with ice have nonzero velocities defined on their margin edges so that
ice can advect out of those cells and the ice margin can advance.  These
boundary cells can either be floating or grounded, and the scalar values
inserted into the FEM mesh depends on which type is being applied.  The
logic for which type gets applied at each boundary node depends on
conditions at the neighboring nodes and edges.

I discovered that when grounded boundaries were applied, there was no
check to ensure that bedTopography was actually above sea level, and in
the complex terrain of AIS simulations, there were many locations where
grounded boundaries were applied below sea level.  Specifically, an ice
thickness of 1 m was applied to a bed elevation below sea level, leading
to an unphysical surface elevation below sea level (and associated high
slopes and other issues).

This commit adds logic to the Interface code to check that grounded
boundaries are above sea level, and, if not, modify them to be a
floating boundary with a thin (~1 m) ice shelf. This is basically a
grounded glacier margin that thins to 0 at the coast.

Before this fix AIS20km runs with GLP were giving maximum velocities of
~300,000 m/yr at these weird locations.  After the fix, those weird
places look normal, and the max speed is ~6000 m/yr.