Force config_basal_mass_bal_float=none to give 0 melt
The logic for config_basal_mass_bal_float was a bit incomplete, such
that the option 'none' did nothing, rather than giving zero melt.  This
meant that if floatingBasalMassBal was read in from a file, that value
was used, making 'none' and 'file' identical.  Because
floatingBasalMassBal is now a restart variable, the 'none' option no
longer worked correctly for MISMIP+ Ice2ra - in that situation the new
run, which should have no melt, inherited the melt rate from the
previous run.

This commit improves the logic so that the 'none' option sets the
floatingBasalMassBal field to zero.