Only create the bdExtensionMap once
The logic for creating the bdExtensionMap was duplicated in the
import2dfields and import2dfieldsObservations, which would potentially
lead to problems if the mapping logic changed (and was already different
in the previous commits after I rebased).  This commit has only
import2dfields create the map and then it gets passed into the second
routine.