Don't write output at initial time of restart
If output was supposed to occur at this time, it should have been
written on the previous run.  Writing it now could/would clobber that
pre-existing data, and not all the fields are filled out on initial
time, which can leave some "holes" in the output file.

This commit skips writing of output and analysis members on the initial
time of a restart.  Note that for debugging inexact restarts it can be
useful to enable these output operations.