Fixes momentum flux in gustiness parameter
This bug and its fix is found by Po-Lun Ma. The original code has a
bug,which makes the gustiness effect on momentum flux smaller than
what it should be. The effects of the bug on clouds and precipitation
are small, but has an important effect on the westerly anomalies in
tropical Pacific which is relevant to ENSO. The bug fix enhances the
momentum flux in the tropics, weakening the prevailing easterlies so
that the cold tongue won't extend too far into the west. The largest
signal comes from TWP, so it might also improve WWB to help ENSO.

[BFB] - Bit-For-Bit