Move SGH solve after prepare_advection so dt is defined
The prepare_advection routine sets the master adaptive time step, and so
should be the first thing run on each time step.  This commit moves it
to the beginning (SGH solve previously occurred first).  Without this
change, the SGH model could think that the master dt was 0, and it would
not run.