Limit temperature to PMP for 'sfc_air_tempature' temp init option
The option config_temperature_init=='sfc_air_temperature' previously set the
initial ice temperature to the minimum of the surface air temperature
and 273.15 K.  However, this led to enormous melt rates on the first
time step when the thermal solver encountered ice temperature well above
the pressure-melt temperature at depth.

This commit modifies that initialization option to instead take the
minimum of the surface air temperature and the PMP temperature at every
depth.