Move SGH solve after thermal solve within the timestep
It's not clear where the best place to call SGH should be.
But it seems sensible to put it after thermal evolution is complete to get
updated basal melting source term.
Also seems (might be?) better to put it after geometry evolution.
We definietly want it before the velocity solve since the hydro model can
control the velo basal b.c.

This commit moves it after themal and evolution solves, but before
velocity.