Add groundedBasalMassBal as restart var when SGH model is active
This field is needed as restart when config_SGH_basal_melt='thermal',
but adding it always.  Conversely, basalMeltInput is not needed as a
restart variable when config_SGH_basal_melt is NOT 'thermal', but
leaving it in as well.  Better to have unnecessary fields than missing
fields, and logic to control those details gets complicated.