Add update to geometry to restore_calving and velocity solve
These updates are needed for the GIS20km test case in ACME to run
successfully - the reworking of the diagnostic variables routines
changed answers for that test case in ACME, where the restore_calving
option is on.

This commit also fixes bit-reproducibility for that test, which appears
to have been broken prior to the reworking of the diagnostic variables
routines.

These changes are answer-changing, but correct.