Update MPASLI to fix bit reproducibility
The previous update of MPASLI in this branch brought in changes that
were not BFB for the configuration used in ACME.  In fixing that bug, I
also discovered that the configuration in ACME was not bit-reproducible.
So this update also fixes that bug.  However, that fix makes this update
answer-changing.  Differences are small, and the new results are
identical across different processor counts.

I also update the namelist option to force the Albany mesh to be rebuilt
on every time step.  This is needed to allow Albany runs to be
BFB-restartable until a fix in Albany is brought into ACME.

[Non-BFB]