Delete redundant betaSolve=beta line when no GLP is in use
We may be  using the effective pressure from the hydrology model to
calculate betaSolve. If so, we want to avoid clobbering betaSolve by
reassigning beta to it under grounded ice.  betaSolve=beta is the
default operation that occurs before the optional application of effective
pressure so it is not necessary to repeat it anyway.