Check for NaNs and INFs in input to shr_reprosum_calc
In recent development NaNs and INFs are often first
identified when passed to the shr_reprosum_calc routine,
where they either lead to segmentation faults or to
very slow performance (due to the interaction of NaNs and
INFs with the reproducible sum logic). To more readily
identify this error condition (and to prevent false
attributions of error to shr_reprosum_calc), the input array
is checked for the presence of NaNs and INFs. If found,
an appropriate error message is output and the job is
terminated.

[BFB]

*********1*********2*********3*********4*********5*********6*********7**
Longer commit message body describing the commit.
Can contain lists as follows:
	* Item 1
	* Item 2
	* Item 3

A good commit message should be written like an email, a subject
followed by a blank line, followed by a more descriptive body.

Can also contain a tag at the bottom describing what type of commit this is.
[BFB] - Bit-For-Bit
[FCC] - Flag Climate Changing
[Non-BFB] - Non Bit-For-Bit
[CC] - Climate Changing
[NML] - Namelist Changing

See confluence for a more detailed description about these tags.