Separate min/max tracer into hor. and vert. variables.
This commit is bit for bit with the previous commit, which was Todd's
"two line fix" from Aug 12 2017.  Here I separate horizontal and
vertical min/max to make it clear.

WARNING: In this commit, FCT scale factor is only chosen from horizontal
min and max.  This is incorrect, but better than only limiting in
the vertical, which is what occurred before 170812.