Initial commit of branch for coupling with GCAM
This is my initial commit of the add-gcam branch, which includes the code I
swiped from the iESM giac subdirectory from the lnd component.  The idea
here is that I'm going to try and reuse as much stuff as I can from iESM,
so I'm dragging all that code along with this commit.

There are no changes to the existing ACME code yet, and we are still in
early development (i.e. there is no reason to think that this branch will
compile, much less run.)