Allow time series stats to use _restart_stream = 'none'
Before this commit, a restart stream was required for every time series
stat namelist.  The user should be able to specify 'none' when normal
model restarts correspond with time averaging intervals.  For example,
the most common case in ACME is:
- Yearly (or monthly) model restarts
- Monthly averaged files for AM_timeSeriesStatsMonthly

In this case, writing a restart file for timeSeriesStatsMonthly is not
needed, because that AM always starts with zeros on a month boundary.
Writing restarts for monthly averaged files takes unnecessary i/o time
and disk space.