Fixed the settings of pgi compiler on Titan
This is a part of PR #1767 submitted by Pat Worley, focusing on the
changes of the settings of pgi compiler in Titan.

As stated in PR #1767:
"Changes including removing the dependence on the version of pgi/17.5.0
installed in Dave Norton's directories, moving instead
to the now official OLCF installation. Also fixed a typo in the
module switch command for the PGI version used with PGIACC, and
changed the 'pin' flag to 'pinned' for PGIACC, as 'pin' is no
longer legal."

[Non-BFB] - Non Bit-For-Bit