Change the mode of the case_setup call in slr
This is one of the few cases in the system tests where you don't
want to run case_setup in test mode. Test-mode prevents the case.test
file from being regenerated. Since the call is made before the call to
build_indv, the TestStatus file should be in good shape.