Allow queue selection by node count instead of task count
All the HPCs care about is how many nodes you request, so it makes
more sense to do selection by nodes instead of tasks. Also, the
mapping of tasks to nodes depends on the number of threads, so
task count alone is not a detemistic way of know how many nodes
you'll need.

1) Chama and skybridge to select queue based on node count
2) Change job-based task_count override to a node_count override
3) XSD changes to support these changes
4) Batch system setup must happen after initialize_derived_attributes (computes num_nodes)
5) Lots of changes to env_batch to support nodemin, nodemax