Force mass_gradient method weight to be positive.
This change has no effect if MG2's mass_gradient method is not used, or
if the alpha and beta parameters to that method are both set to 1 or
less. However, this change is necessary to maintain a positive value for
precipitation fraction in the presence of large gradients when alpha or
beta are larger than 1.

[BFB]