Avoid divide-by-zero in MG2 ice sublimation.
The variable `nsubi` in MG2 is calculated by dividing by `qi`. This
means that we need to check for (near-)zero values to avoid triggering a
floating-point exception. It appears that the resulting NaN values are
filtered out by limiters, so this commit may not end up changing
answers, but that is unclear at this time.

[non-BFB]