Replace pio_swapm call in compute_counts with pio_N2M
High MPI overhead is incurred on some systems the
first time that two processes communicate. While performance of
the pio_swapm command is excellent during typical invocations,
its cost can be as much as twice that of the special purpose routine
pio_N2M when called the first time (for reasons that are not
completely clear). As pio_N2M uses the same communication pattern
as pio_swapm for both the M-to-N scatters and N-to-M gathers used in
PIO, subsequent calls to pio_swapm do not incur the MPI start-up
overhead once pio_N2M is called from within the compute_counts
routine.

[BFB]