Fix broken builds with the PGI compilers
The PGI compilers apparently need some module files from the ESMF timekeeping
library when compiling code in the physics_wrf directory, and they also need
the 'esmf_time' library when linking the build_tables utility program.
Presumably, this is due to some indirect use of the ESMF modules that came
with the use of the new logging system.